refactor: rearrange code
